+++ lib/extensions/file_extension.dart
@@ -38,4 +38,7 @@ extension FileExtension on File {
   String getSizeLabel({int asFixed = 2}) {
     return statSync().size.toDouble().toFileSizeLabel(asFixed: asFixed);
   }
+  String toFileSizeLabel({int asFixed = 2}) {
+    return statSync().size.toDouble().toFileSizeLabel(asFixed: asFixed);
+  }
 }