+++ lib/extensions/int_extension.dart
@@ -4,4 +4,7 @@ extension IntExtension on int {
   String getSizeLabel({int asFixed = 2}) {
     return toDouble().toFileSizeLabel(asFixed: asFixed);
   }
+  String toFileSizeLabel({int asFixed = 2}) {
+    return toDouble().toFileSizeLabel(asFixed: asFixed);
+  }
 }