+++ lib/interfaces/my_method_channel.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
-
-class MyMethodChannel {
-  @visibleForTesting
-  final channel = const MethodChannel('than_pkg');
-}