+++ lib/interfaces/than_pkg_method_channel.dart
@@ -6,5 +6,5 @@ import 'than_pkg_platform_interface.dart';
 class ThanPkgMethodChannel extends ThanPkgPlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
-  final methodChannel = const MethodChannel('than_pkg');
+  static final methodChannel = const MethodChannel('than_pkg');
 }