+++ lib/t_database/t_database_listener.dart
@@ -1,3 +1,0 @@
-mixin TDatabaseListener {
-  void onDatabaseChanged();
-}