+++ lib/t_database/t_json_database.dart
@@ -1,64 +1,40 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'dart:convert';
 
-import 'data_io.dart';
-import 't_database_listener.dart';
+import 'package:than_pkg/t_database/t_database.dart';
 
-abstract class TJsonDatabase<T> {
+import 'data_io.dart';
+
+abstract class TJsonDatabase<T> extends TDatabase<T> {
   final DataIO io;
-  final String path;
-  TJsonDatabase({required this.path}) : io = JsonIO.instance;
+  TJsonDatabase({required super.root}) : io = JsonIO.instance;
 
   T fromMap(Map<String, dynamic> map);
   Map<String, dynamic> toMap(T value);
-  // ကြိုပေးထား
 
-  Future<List<T>> getAll() async {
-    final json = await io.read(path);
+  @override
+  Future<List<T>> getAll({Map<String, dynamic>? query = const {}}) async {
+    final json = await io.read(root);
     if (json.isEmpty) return [];
     List<dynamic> jsonList = jsonDecode(json);
     return jsonList.map((e) => fromMap(e)).toList();
   }
 
+  @override
   Future<void> add(T value) async {
     final list = await getAll();
     list.add(value);
-    await save(list);
-  }
-
-  Future<void> update(int index, T value) async {
-    final list = await getAll();
-    list[index] = value;
-    await save(list);
-  }
-
-  Future<void> delete(int index) async {
-    final list = await getAll();
-    list.removeAt(index);
+    notify(TDatabaseListenerTypes.add, null);
     await save(list);
   }
 
   Future<void> save(List<T> list, {bool isPretty = true}) async {
     final jsonList = list.map((e) => toMap(e)).toList();
-    await io.write(path, JsonEncoder.withIndent(' ').convert(jsonList));
-    notify();
-  }
-
-  // listener
-  final List<TDatabaseListener> _listeners = [];
-
-  // register/unregister methods
-  void addListener(TDatabaseListener listener) {
-    _listeners.add(listener);
-  }
-
-  void removeListener(TDatabaseListener listener) {
-    _listeners.remove(listener);
-  }
-
-  void notify() {
-    for (final listener in _listeners) {
-      listener.onDatabaseChanged();
+    if (isPretty) {
+      await io.write(root, JsonEncoder.withIndent(' ').convert(jsonList));
+    } else {
+      await io.write(root, jsonEncode(jsonList));
     }
+    notify(TDatabaseListenerTypes.saved, null);
   }
 }